backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
supabase>=2.0.0
anthropic>=0.18.0
praw>=7.7.0
youtube-transcript-api>=0.6.0
scrapetube>=2.5.0
requests>=2.31.0
tidalapi>=0.7.0

